+++ js/selectores.js
@@ -1,0 +1,11 @@
+// Campos del formulario
+export const mascotaInput = document.querySelector('#mascota');
+export const propietarioInput = document.querySelector('#propietario');
+export const telefonoInput = document.querySelector('#telefono');
+export const fechaInput = document.querySelector('#fecha');
+export const horaInput = document.querySelector('#hora');
+export const sintomasInput = document.querySelector('#sintomas');
+
+// UI
+export const formulario = document.querySelector('#nueva-cita');
+export const contenedorCitas = document.querySelector('#citas');
